Split callback tests to their own source file
I intend to do some work to clean up callback routine behaviors, it will be easier to do this with a test file dedicated to testing them.
